+++ CHANGELOG.md
@@ -14,6 +14,10 @@ as `## [X.Y.Z] - YYYY-MM-DD` and setting `meson.build` to the same version
 
 ## [Unreleased]
 
+### Added
+
+- Connection dialog switch to use the passphrased identity; unprotected keys stay available for scripts
+
 ## [0.1.3] - 2026-08-19
 
 ### Added